refactor: 기존 Product, ProductDao 제거
