--- src/main/java/kitchenpos/domain/Product.java
@@ -1,47 +1,0 @@
-package kitchenpos.domain;
-
-import java.math.BigDecimal;
-
-public class Product {
-    private Long id;
-    private String name;
-    private BigDecimal price;
-
-    public Product() {
-    }
-
-    public Product(String name, BigDecimal price) {
-        this.name = name;
-        this.price = price;
-    }
-
-    public Product(Long id, String name, BigDecimal price) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(final Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(final String name) {
-        this.name = name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
-    }
-}
